+++ src/main/java/scanner/AbstractRunner.java
@@ -1,0 +1,14 @@
+package scanner;
+
+import com.google.common.base.Stopwatch;
+
+/**
+ * Common abstract runner.
+ *
+ * @author inkarnadin
+ */
+public abstract class AbstractRunner {
+
+    protected Stopwatch timer = Stopwatch.createStarted();
+
+}