+++ src/main/java/scanner/Runner.java
@@ -1,8 +1,0 @@
-package scanner;
-
-/**
- * Common runner interface.
- *
- * @author inkarnadin
- */
-public interface Runner {}