+++ src/main/java/scanner/stat/TimeStatItem.java
@@ -1,0 +1,15 @@
+package scanner.stat;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum TimeStatItem {
+
+    TOTAL_TIME,
+    EXPECTED_TIME,
+    TOTAL_SCAN_TIME,
+    TOTAL_BRUTE_TIME,
+    AVG_SCAN_TIME,
+    AVG_BRUTE_TIME;
+
+}